+++ yt/server/scheduler/map_controller.cpp
@@ -340,7 +340,7 @@ private:
 
     void InitJobSpecTemplate()
     {
-        JobSpecTemplate.set_type(EJobType::Map);
+        JobSpecTemplate.set_type(static_cast<int>(EJobType::Map));
         auto* schedulerJobSpecExt = JobSpecTemplate.MutableExtension(TSchedulerJobSpecExt::scheduler_job_spec_ext);
         schedulerJobSpecExt->set_lfalloc_buffer_size(GetLFAllocBufferSize());
